Move text completion settings to respective module
